Move AsyncResult syntax to syntax package
