+++ core/src/main/scala/janstenpickle/vault/core/Secrets.scala
@@ -1,6 +1,7 @@
 package janstenpickle.vault.core
 
 import com.ning.http.client.Response
+import janstenpickle.scala.syntax.AsyncResultSyntax._
 import janstenpickle.scala.syntax.SyntaxRequest._
 import janstenpickle.scala.syntax.ResponseSyntax._
 import janstenpickle.scala.syntax.VaultConfigSyntax._
